+++ Configuration/Services.php
@@ -14,12 +14,12 @@ namespace Brotkrueml\Schema;
 use Brotkrueml\Schema\Cache\PagesCacheService;
 use Brotkrueml\Schema\Configuration\Configuration;
 use Brotkrueml\Schema\Configuration\ConfigurationProvider;
+use Brotkrueml\Schema\Core\AdditionalPropertiesInterface;
 use Brotkrueml\Schema\Core\Model\TypeInterface;
+use Brotkrueml\Schema\DependencyInjection\AdditionalPropertiesPass;
 use Brotkrueml\Schema\DependencyInjection\TypeProviderPass;
 use Brotkrueml\Schema\EventListener\AddBreadcrumbList;
 use Brotkrueml\Schema\EventListener\AddWebPageType;
-use Brotkrueml\Schema\EventListener\RegisterRemovedTypePropertiesForPhysician;
-use Brotkrueml\Schema\EventListener\RegisterTypePropertiesMovedFromOfficialToPending;
 use Brotkrueml\Schema\Hooks\PageRenderer\SchemaMarkupInjection;
 use Brotkrueml\Schema\Lowlevel\ConfigurationModuleProvider\Types;
 use Brotkrueml\Schema\Manager\SchemaManager;
@@ -32,6 +32,9 @@ use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
 return static function (ContainerConfigurator $configurator, ContainerBuilder $builder): void {
     $builder->registerForAutoconfiguration(TypeInterface::class)->addTag('tx_schema.type');
     $builder->addCompilerPass(new TypeProviderPass('tx_schema.type'));
+
+    $builder->registerForAutoconfiguration(AdditionalPropertiesInterface::class)->addTag('tx_schema.additional_properties');
+    $builder->addCompilerPass(new AdditionalPropertiesPass('tx_schema.additional_properties'));
 
     $services = $configurator->services();
     $services->defaults()
@@ -73,16 +76,6 @@ return static function (ContainerConfigurator $configurator, ContainerBuilder $b
         ->arg('$configuration', service('schema.configuration'))
         ->tag('event.listener', [
             'identifier' => 'ext-schema/addWebPageType',
-        ]);
-
-    $services->set(RegisterTypePropertiesMovedFromOfficialToPending::class)
-        ->tag('event.listener', [
-            'identifier' => 'ext-schema/registerTypePropertiesMovedFromOfficialToPending',
-        ]);
-
-    $services->set(RegisterRemovedTypePropertiesForPhysician::class)
-        ->tag('event.listener', [
-            'identifier' => 'ext-schema/registerRemovedTypePropertiesForPhysician',
         ]);
 
     $services->set(SchemaContentObject::class)